Stomp longs across bitness (#2)
* Change instances of long to int for:
- CSHA1
- bf_read/write
- basetypes.h
- CBitVecAccessor
- CBaseHandle
- UTIL_ParticleEffect

* Adjust long usage in threadtools
Recompile tier1.a
 - Fix incorrect clang version outputted for 10 and up
Change CThreadFastMutex::m_ownerID to ThreadId_t
 - tier0 exports CThreadFastMutex::Lock(), first param is threadId and second param is sleep time. On 64-bit, threadId is a long long, but is a long on 32-bit.
 - CKeyValuesGrowableStringTable::GetSymbolForString uses a CThreadFastMutex, comparing where m_depth member is set is off by 4, implying that CThreadFastMutex::m_ownerID's size was changed across arch's
 - With the diff. in m_ownerID's size, use of interlock intrinsics need to be adjusted
Redo ThreadInterlocked* extern's
 - tier0 only exports 64-bit interlocks on 32-bit; none of the interlocks are exported on linux32, linux64, or win64 and are inlined (intrinsics)

* Recompile tier1.lib for 🪟
Fix size_t warnings
Update LZMA to 2019-02-02
- 2013 LZMA is uncompilable on 64-bit and honestly I CBA to figure out the exact version used by the beta, so just use the latest one. This 17 year old game is getting Vulcan rendering for goodness sake.
Fix X64BITS being used in CUtlBuffer::GetPtr
 - Supposed to be PLATFORM_64BITS
Fix warnings about long being truncated to int
Use V_strlen over regular strlen and other functions using int's

* Recompile mathlib.lib for 🪟
Use SSE2 instruction set
 - SSE2 is the absolute minimum required for x86-64 and is enabled by default on 64-bit, so no need to manually specify to use the architecture (https://learn.microsoft.com/en-us/cpp/c-runtime-library/reference/set-sse2-enable?view=msvc-140)

* Recompile tier1.a and mathlib.a against SteamRT 3.0 (Sniper)

* Fix x64 detection in mathlib

* Add Mix64HashFunctor implementation

+++ game/server/gameinterface.cpp
@@ -3388,7 +3388,7 @@ void MessageWriteEHandle( CBaseEntity *pEntity )
 	if (!g_pMsgBuffer)
 		Error( "WriteEHandle called with no active message\n" );
 
-	long iEncodedEHandle;
+	int iEncodedEHandle;
 	
 	if( pEntity )
 	{
