+++ game/server/util.cpp
@@ -1341,7 +1341,7 @@ void UTIL_SetOrigin( CBaseEntity *entity, const Vector &vecOrigin, bool bFireTri
 }
 
 
-void UTIL_ParticleEffect( const Vector &vecOrigin, const Vector &vecDirection, ULONG ulColor, ULONG ulCount )
+void UTIL_ParticleEffect( const Vector &vecOrigin, const Vector &vecDirection, unsigned int ulColor, unsigned int ulCount )
 {
 	Msg( "UTIL_ParticleEffect:  Disabled\n" );
 }
